Split frame_diff.cpp out from frame.cpp
